requirements.txt
flask==3.0.0
flask-cors==4.0.0
boto3==1.35.0
requests==2.32.5
google-generativeai==0.8.5
pinecone-client==3.0.0
opencv-python==4.10.0
torch==2.6.1
torchvision==0.21.1
transformers==4.47.1
pillow==11.1.0
tqdm==4.67.1